requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
openai
python-dotenv